Split actions into books and filter and add actionTypes file
