+++ src/js/actions/filter.js
@@ -1,0 +1,6 @@
+import { CHANGE_FILTER } from './actionTypes';
+
+export const changeFilter = filterText => ({
+  type: CHANGE_FILTER,
+  filterText,
+});
