+++ src/js/containers/BookList.js
@@ -1,19 +1,20 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import { removeBook, changeFilters } from '../actions';
+import { removeBook } from '../actions/books';
+import { changeFilter } from '../actions/filter';
 import Book from '../components/Book';
 import CategoryFilter from '../components/CategoryFilter';
 
 export const BookList = (props) => {
   const {
-    books, changeFilters, filter, removeBook,
+    books, changeFilter, filter, removeBook,
   } = props;
   return (
     <div>
       <CategoryFilter
         filter={filter}
-        handleChange={optionValue => changeFilters(optionValue)}
+        handleChange={optionValue => changeFilter(optionValue)}
       />
       <div>
         {books.map(book => (
@@ -31,12 +32,12 @@ const mapStateToProps = state => ({
 
 BookList.propTypes = {
   books: PropTypes.instanceOf(Object).isRequired,
-  changeFilters: PropTypes.func.isRequired,
+  changeFilter: PropTypes.func.isRequired,
   filter: PropTypes.string.isRequired,
   removeBook: PropTypes.func.isRequired,
 };
 
 export default connect(
   mapStateToProps,
-  { removeBook, changeFilters },
+  { removeBook, changeFilter },
 )(BookList);
